requirements.txt
dj-database-url==0.4.2
Flask==0.12.2
Flask-SQLAlchemy==2.2
Flask-WTF==0.14.2
python-decouple==3.0
requests==2.17.3
